+++ include/apxchol/solver/sptrsv/cuda_levelset.h
@@ -45,8 +45,12 @@ void levelset_solve(cudaStream_t stream,
 /// divides by is the fp32 `diag[i]` (fp32(L_ii / s_i), plus the column's
 /// rounding residual under the default diag_comp) -- and the row's rhs is
 /// scaled once by `in_scale[i]` if `in_scale` != nullptr (the back solve
-/// passes inv_scale^2 = fp32(1/s_i)^2, folding D^-2 into its input; the
-/// forward solve passes nullptr):
+/// passes inv_scale^2 = fp32(1/s_i)^2 held in DOUBLE -- r_i^2 exact; the
+/// product rhs * r_i^2 is formed in double and narrowed once, because the
+/// pre-squared value overflows fp32 for s_i < ~5.4e-20 while the product
+/// itself is small (the CPU FP16_SCALED contract, omp.h back_dir::rhs,
+/// computes the very same double product) -- folding D^-2 into its input;
+/// the forward solve passes nullptr):
 ///   out[i] = (rhs[i] * in_scale[i] - sum_{j != i} widen(vals16[p]) * out[j]) / diag[i].
 /// Every product is formed and accumulated in sptrsv_gpu_value_t (float on
 /// the default fp32 build) after the half -> float widen; the fixed warp-
@@ -54,7 +58,7 @@ void levelset_solve(cudaStream_t stream,
 /// per-level launch structure as levelset_solve.
 void levelset_solve_fp16(cudaStream_t stream,
                          const int* rowptr, const int* colidx, const unsigned short* vals16,
-                         const float* diag, const float* in_scale,
+                         const float* diag, const double* in_scale,
                          const int* row_order, const int* level_ptr, int num_levels,
                          const sptrsv_gpu_value_t* rhs, sptrsv_gpu_value_t* out);
